Renaming, one class per file, documentation, minor edits
