+++ src/Random/Utility.h
@@ -1,0 +1,279 @@
+// -*- C++ -*-
+/***************************************************************************
+ *
+ * The IPPL Framework
+ *
+ * This program was prepared by PSI.
+ * All rights in the program are reserved by PSI.
+ * Neither PSI nor the author(s)
+ * makes any warranty, express or implied, or assumes any liability or
+ * responsibility for the use of this software
+ *
+ * Visit www.amas.web.psi for more details
+ *
+ ***************************************************************************/
+
+#ifndef IPPL_RANDOM_UTILITY_H
+#define IPPL_RANDOM_UTILITY_H
+
+#include <Kokkos_MathematicalConstants.hpp>
+#include <Kokkos_MathematicalFunctions.hpp>
+#include <Kokkos_Random.hpp>
+#include "Types/ViewTypes.h"
+
+namespace ippl {
+  namespace random {
+  
+    /*!
+     * @struct randn
+     * @brief Functor to generate random numbers from a normal distribution.
+     *
+     * This functor can be used to generates random numbers from a normal distribution with
+     * mean 0 and standard deviation 1.
+     *
+     * @tparam T Data type of the random numbers.
+     * @tparam GeneratorPool Type of the random number generator pool.
+     * @tparam Dim Dimensionality of the random numbers.
+    */
+    template <typename T, class GeneratorPool, unsigned Dim>
+    struct randn {
+      using view_type  = typename ippl::detail::ViewType<T, 1>::view_type;
+      // Output View for the random numbers
+      view_type v;
+
+      // The GeneratorPool
+      GeneratorPool rand_pool;
+
+      /*!
+       * @brief Constructor for the randn functor.
+       *
+       * @param v_ Output view for the random numbers.
+       * @param rand_pool_ The random number generator pool.
+      */
+      randn(view_type v_, GeneratorPool rand_pool_)
+          : v(v_)
+          , rand_pool(rand_pool_) {}
+
+      /*!
+       * @brief Operator to generate random numbers.
+       *
+       * @param i Index for the random numbers.
+      */
+      KOKKOS_INLINE_FUNCTION void operator()(const size_t i) const {
+          // Get a random number state from the pool for the active thread
+          typename GeneratorPool::generator_type rand_gen = rand_pool.get_state();
+
+          for (unsigned d = 0; d < Dim; ++d) {
+              v(i)[d] = rand_gen.normal(0.0, 1.0);
+          }
+
+          // Give the state back, which will allow another thread to acquire it
+          rand_pool.free_state(rand_gen);
+      }
+    };
+
+    /*!
+     * @struct randu
+     * @brief Functor to generate random numbers from a uniform distribution.
+     *
+     * This functor generates random numbers from a uniform distribution in the
+     * range [0, 1].
+     *
+     * @tparam T Data type of the random numbers.
+     * @tparam GeneratorPool Type of the random number generator pool.
+     * @tparam Dim Dimensionality of the random numbers.
+    */
+    template <typename T, class GeneratorPool, unsigned Dim>
+    struct randu {
+      using view_type  = typename ippl::detail::ViewType<T, 1>::view_type;
+      // Output View for the random numbers
+      view_type v;
+
+      // The GeneratorPool
+      GeneratorPool rand_pool;
+
+      /*!
+       * @brief Constructor for the randu functor.
+       *
+       * @param v_ Output view for the random numbers.
+       * @param rand_pool_ The random number generator pool.
+      */
+      randu(view_type v_, GeneratorPool rand_pool_)
+        : v(v_)
+        , rand_pool(rand_pool_) {}
+
+      /*!
+       * @brief Operator to generate random numbers.
+       *
+       * @param i Index for the random numbers.
+      */
+      KOKKOS_INLINE_FUNCTION void operator()(const size_t i) const {
+        // Get a random number state from the pool for the active thread
+        typename GeneratorPool::generator_type rand_gen = rand_pool.get_state();
+
+        for (unsigned d = 0; d < Dim; ++d) {
+            v(i)[d] = rand_gen.drand(0.0, 1.0);
+        }
+
+	  // Give the state back, which will allow another thread to acquire it
+        rand_pool.free_state(rand_gen);
+      }
+    };
+  
+    /*!
+     * @brief Calculate the cumulative distribution function (CDF) for a normal distribution.
+     *
+     * @param x Input value.
+     * @param mean Mean of the distribution.
+     * @param stddev Standard deviation of the distribution.
+     * @return The CDF value.
+    */
+    template<typename T>
+    KOKKOS_FUNCTION T normal_cdf_func(T x, T mean, T stddev) {
+      return 0.5 * (1 + Kokkos::erf((x - mean) / (stddev * Kokkos::sqrt(2.0))));
+    }
+  
+    /*!
+     * @brief Calculate the probability density function (PDF) for a normal distribution.
+     *
+     * @param x Input value.
+     * @param mean Mean of the distribution.
+     * @param stddev Standard deviation of the distribution.
+     * @return The PDF value.
+    */
+    template<typename T>
+    KOKKOS_FUNCTION T normal_pdf_func(T x, T mean, T stddev) {
+      const T pi = Kokkos::numbers::pi_v<T>;
+      return (1.0 / (stddev * Kokkos::sqrt(2 * pi))) * Kokkos::exp(-(x - mean) * (x - mean) / (2 * stddev * stddev));
+    }
+  
+    /*!
+     * @brief An estimator for the initial guess that is used in Newton-Raphson method of Inverste Transfrom Sampling
+     *
+     * @param x Input value.
+     * @param mean Mean of the distribution.
+     * @param stddev Standard deviation of the distribution.
+     * @return The estimate value.
+    */
+    template<typename T>
+    KOKKOS_FUNCTION T normal_estimate_func(T u, T mean, T stddev) {
+      const T pi = Kokkos::numbers::pi_v<T>;
+      return (Kokkos::sqrt(pi / 2.0) * (2.0 * u - 1.0)) * stddev + mean;
+    }
+  
+    template<typename T>
+    KOKKOS_FUNCTION T uniform_cdf_func(T x){
+      return x;
+    }
+  
+    template<typename T>
+    KOKKOS_FUNCTION T uniform_pdf_func(){
+      return 1.;
+    }
+  
+    template<typename T>
+    KOKKOS_FUNCTION T uniform_estimate_func(T u){
+      return u;
+    }
+  
+    /*!
+     * @struct normal_cdf
+     * @brief Functor to calculate the cumulative distribution function (CDF) for a normal distribution.
+     *
+     * This functor calculates the CDF for a normal distribution in a specific dimension 'd'.
+     *
+     * @tparam T Data type for the input value 'x'.
+     * @tparam Dim Dimensionality of the distribution.
+    */
+    template <typename T, unsigned Dim>
+    struct normal_cdf{
+      KOKKOS_INLINE_FUNCTION double operator()(T x, unsigned int d, const T *params) const {
+              T mean = params[d*Dim + 0];
+              T stddev = params[d*Dim + 1];
+              return ippl::random::normal_cdf_func<T>(x, mean, stddev);
+      }
+    };
+  
+    /*!
+     * @struct normal_pdf
+     * @brief Functor to calculate the probability density function (PDF) for a normal distribution.
+     *
+     * This functor calculates the PDF for a normal distribution in a specific dimension 'd'.
+     *
+     * @tparam T Data type for the input value 'x'.
+     * @tparam Dim Dimensionality of the distribution.
+    */
+    template <typename T, unsigned Dim>
+    struct normal_pdf{
+      KOKKOS_INLINE_FUNCTION double operator()(T x, unsigned int d, T const *params) const {
+              T mean = params[d*Dim + 0];
+              T stddev = params[d*Dim + 1];
+              return ippl::random::normal_pdf_func<T>(x, mean, stddev);
+      }
+    };
+  
+    /*!
+     * @struct normal_estimate
+     * @brief Functor to estimate the initial guess for sampling normal distribution.
+     *
+     * This functor estimates the value for a normal distribution in a specific dimension 'd'.
+     *
+     * @tparam T Data type for the input value 'u'.
+     * @tparam Dim Dimensionality of the distribution.
+    */
+    template <typename T, unsigned Dim>
+    struct normal_estimate{
+      KOKKOS_INLINE_FUNCTION double operator()(T u, unsigned int d,  T const *params) const {
+              T mean = params[d*Dim + 0];
+              T stddev = params[d*Dim + 1];
+              return ippl::random::normal_estimate_func<T>(u, mean, stddev);
+      }
+    };
+  
+    namespace detail{
+      /*!
+       * @struct NewtonRaphson
+       * @brief Functor for solving equations using the Newton-Raphson method.
+       *
+       * In particular, find the root x of the equation dist.obj(x, u)= 0 for a given u using Newton-Raphson method.
+       *
+       * @tparam T Data type for the equation variables.
+      */
+      template <typename T>
+      struct NewtonRaphson {
+          KOKKOS_FUNCTION
+          NewtonRaphson() = default;
+
+          KOKKOS_FUNCTION
+          ~NewtonRaphson() = default;
+
+          /*!
+           * @brief Solve an equation using the Newton-Raphson method.
+           *
+           * This function iteratively solves an equation of the form "cdf(x) - u = 0"
+           * for a given sample `u` using the Newton-Raphson method.
+           *
+           * @tparam Distribution Type of the distribution providing the objective function and its derivative.
+           * @param dist Distribution object providing objective function cdf(x)-u and its derivative.
+           * @param d Dimension index.
+           * @param x Variable to solve for (initial guess and final solution).
+           * @param u Given sample from a uniform distribution [0, 1].
+           * @param atol Absolute tolerance for convergence (default: 1.0e-12).
+           * @param max_iter Maximum number of iterations (default: 20).
+          */
+          template <class Distribution>
+          KOKKOS_INLINE_FUNCTION void solve(Distribution dist, int d, T& x, T& u, T atol = 1.0e-12,
+                                            unsigned int max_iter = 20) {
+              unsigned int iter = 0;
+              while (iter < max_iter && Kokkos::fabs(dist.obj_func(x, d, u)) > atol) {
+                  // Find x, such that "cdf(x) - u = 0" for a given sample of u~uniform(0,1)
+                  x = x - (dist.obj_func(x, d, u) / dist.der_obj_func(x, d));
+                  iter += 1;
+              }
+          }
+      };
+    }  // name space detial
+  }  // namespace random
+}  // namespace ippl
+
+#endif
